Added Vertex Raymarching
Added vertex raymarching. Unfortunately, Managed to roughly double the number of effective marches with very little performance overhead. I use "double" loosely because the vertex marches are clearly not of the same quality and are of very little use besides getting large sections of the screen close to their nearest geometry.

Its sort of like local dimming, and the vertex shader is only good enough to give use 256 zones.

Next up is to recolor everything so that it looks good. Especially so that non rendering objects look like fog or ambient occlusion.